+++ Examen_U1_POO_CarlosPineda/Examen_U1_POO_CarlosPineda/Dtos/Productos/ProductoDto.cs
@@ -6,7 +6,7 @@
 
         public string Name { get; set; } 
 
-        public double Cantidad { get; set; }
+        public int Cantidad { get; set; }
 
         public double Precio { get; set; }  
     }
